+++ crates/storage-ops/src/pruning/ledger/l2_blocks.rs
