+++ crates/storage-ops/src/pruning/ledger/slots.rs
