+++ crates/storage-ops/src/pruning/state.rs
