--- crates/storage-ops/src/pruning/types.rs
@@ -1,1 +1,0 @@
-
