--- crates/storage-ops/src/rollback/components/mod.rs
@@ -1,7 +1,0 @@
-mod ledger_db;
-mod native_db;
-mod state_db;
-
-pub(crate) use ledger_db::*;
-pub(crate) use native_db::*;
-pub(crate) use state_db::*;
