+++ crates/storage-ops/src/rollback/ledger/mod.rs
@@ -1,3 +1,5 @@
 pub(crate) mod commitments;
 pub(crate) mod l2_blocks;
 pub(crate) mod slots;
+pub(crate) mod traits;
+pub(crate) mod types;
