+++ crates/storage-ops/src/rollback/ledger/slots.rs
