+++ crates/storage-ops/src/rollback/native.rs
