+++ crates/storage-ops/src/rollback/state.rs
